[FIX] Tactile sensors: hook signatures, kernel fusion, factoring, file split
Hook signatures (P0):
- 6 _update_current_timestep_data overrides gained ground_truth_data_timeline arg
- ContactProbeSensor._post_process gained timeline + is_measured
- GT ring slot 0 written from every override

Architecture:
- Probe mixins (ProbeSensorMixin, ProbesWithNormalSensorMixin, metadata, func_noised_probe_radius)
  moved from base_sensor.py to dedicated sensors/probe.py
- ProbesWithNormalSensorMixin inherits explicitly from ProbeSensorMixin (drops MRO fragility)
- ContactProbeSensor inherits from ContactDepthProbeSensor (shared kernel/pipeline)
- Common probe-sphere debug draw factored into ProbeSensorMixin._compute_probes_world_pos
- kinematic_tactile.py split into kinematic_tactile.py + point_cloud_tactile.py
- New public Collider.activate_sdf() replaces private _sdf.activate() reach

Hot-path / kernel fusion:
- Dual-radius SDF query in _func_query_contact_depth and _func_query_contact_depth_penetration:
  one SDF lookup yields both GT and noised-radius penetrations
- AABB pre-filter expands by max(gt, noised) so neither branch is silently skipped
- ContactProbe threshold row cached on metadata, eliminating per-step torch.zeros
- Measured-cols scratch tensor lazy-allocated on ProbeSensorMetadataMixin

Conventions:
- getattr removed from _active_envs_mask_tensor (direct attribute access)
- Redundant np.asarray on already-numpy magnitude buffer dropped
- Duplicate __init__ in ProximityTaxelSensor removed
- Sentinel slice(None) replaced by None in SurfaceDistanceProbeSensor
- ASCII section banners removed
- reset() in SurfaceDistanceProbeSensor zeros nearest_positions instead of writing link-local positions
  into a world-frame buffer

Tests / examples:
- tests/test_point_cloud.py deleted; 4 packed FPS tests added to tests/test_misc.py
- test_surface_distance_sensor_box_sphere parameter order aligned with peers
- tactile_sandbox.py procedurally generates the torus mesh via trimesh.creation;
  the 2082-line genesis/assets/meshes/torus.obj is removed
